pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "comfydl"
version = "0.1.7"
authors = [{ name = "ShinChven", email = "shinchven@gmail.com" }]
description = "A robust CLI tool for downloading models for ComfyUI"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
dependencies = ["PyYAML", "requests", "tqdm", "questionary"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

[project.urls]
Repository = "https://github.com/ShinChven/comfydl"

[project.scripts]
comfydl = "comfydl.main:main"

[tool.setuptools]
packages = ["comfydl"]
include-package-data = true

[tool.setuptools.package-data]
